+++ src/arch/mips8/instruction/twoRoneI/AddiInstruction.java
@@ -23,7 +23,7 @@ public class AddiInstruction extends TwoRoneIInstruction {
 	@Override
 	public boolean executeEX() {
 		super.executeEX();
-		super.r1Val = super.r2Val + super.immd;
+		super.r1Val = (long)((int)super.r2Val + (int)super.immd);
 		return true;
 	}
